+++ src/features/machines/depositGenerateAddressMachine.ts
@@ -1,164 +1,137 @@
+import type { PreparationOutput } from "src/services/depositService"
+import type { SupportedChainName } from "src/types/base"
+import type { ChainType } from "src/types/deposit"
+import { assert } from "src/utils/assert"
 import { assign, fromPromise, setup } from "xstate"
-import { logger } from "../../logger"
-import type { ChainType } from "../../types/deposit"
-import type { BlockchainEnum } from "../../types/interfaces"
 
-export type DepositGeneratedDescription = {
-  type: "depositAddressGenerated"
-  depositAddress: string
+export type Context = {
+  userAddress: string | null
+  userChainType: ChainType | null
+  blockchain: SupportedChainName | null
+  preparationOutput: PreparationOutput | null
 }
-
-type Context = {
-  userAddress: string
-  userChainType: ChainType
-  chain: BlockchainEnum | null
-  depositAddress: string | null
-  error: null | {
-    tag: "err"
-    value: {
-      reason: "ERR_GENERATING_ADDRESS"
-      error: Error | null
-    }
-  }
-}
-
-type Events = {
-  type: "INPUT"
-  userAddress: string
-  defuseAssetId: string
-}
-
-type Input = {
-  userAddress: string
-  userChainType: ChainType
-  chain: BlockchainEnum
-}
-
-export type Output =
-  | NonNullable<Context["error"]>
-  | {
-      tag: "ok"
-      value: {
-        depositAddress: string
-        depositDescription: DepositGeneratedDescription
-      }
-    }
 
 export const depositGenerateAddressMachine = setup({
   types: {
     context: {} as Context,
-    events: {} as Events,
-    input: {} as Input,
-    output: {} as Output,
+    events: {} as {
+      type: "REQUEST_GENERATE_ADDRESS"
+      params: NonNullable<
+        Pick<Context, "userAddress" | "userChainType" | "blockchain">
+      >
+    },
   },
   actors: {
     generateDepositAddress: fromPromise(
-      async (_: { input: Input }): Promise<string> => {
+      async (_: {
+        input: {
+          userAddress: string
+          userChainType: ChainType
+          blockchain: SupportedChainName
+        }
+      }): Promise<string> => {
         throw new Error("not implemented")
       }
     ),
   },
   actions: {
-    setError: assign({
-      error: (_, error: NonNullable<Context["error"]>["value"]) => ({
-        tag: "err" as const,
-        value: error,
-      }),
+    setInputParams: assign(({ event }) => {
+      return {
+        userAddress: event.params.userAddress,
+        userChainType: event.params.userChainType,
+        blockchain: event.params.blockchain,
+      }
     }),
-    logError: (_, params: { error: unknown }) => {
-      logger.error(params.error)
+    resetPreparationOutput: assign(() => {
+      return {
+        preparationOutput: null,
+      }
+    }),
+  },
+  guards: {
+    isInputSufficient: ({ event }) => {
+      if (
+        event.params.blockchain === "near" ||
+        event.params.blockchain === "turbochain" ||
+        event.params.blockchain === "aurora"
+      ) {
+        return false
+      }
+      return (
+        event.params.userAddress != null &&
+        event.params.userChainType != null &&
+        event.params.blockchain != null
+      )
     },
   },
 }).createMachine({
-  /** @xstate-layout N4IgpgJg5mDOIC5QTABwPawJYBcC0MAdmAE4CGOYeZEEJcsAdFhADZgDEAkgHIAKAVQAqAbQAMAXUSgM2HFnSFpIAB6IAjAGYALIzEA2ABz7NAdgBMO-WM3qANCACeiTWMb6tpgJznD500Y2vgC+wQ4osrgEYMTklNS09LBMRKQUWIRQHBCKYMyEAG7oANZ5EZhRqXFUNHQMjFXpmQgZRQDG6YriEt3KkfKKymoI+tqGjNpill76XgCstgb2Toj6jF5i6luGfl6GYvNjoeFoFfiN8bVJKTFp8pkcpCToJIyorBQAZi8AtozlcmisQoNUS9QuGSgLUK6A6A0I3V6SBA-QUSmRw1GXj0phshjGcyM+M0DmcCC8mkYhlM6jEc0sk3MXj2oTCIEI6BQ8GRAMqt2qCTqyT6ZzRQ0QeHMazEpjmhgphNM+3lAVJEvUzMYcz2xgpMwMpmOIF5535IMF12YbDAIrkYoxiHM6nW5m0tNGmidXlprrVCHUpnczOpPlcmk0s3DRpNQLuoKFN2B9ygttw9tAwyZunp0y82m9vh2pj96l0OiCcv8AVmsujp0BF3jloA4vywJQIKn4eLyTT1tpZh4A6M5V4-f5GOp-PoLGJtJo9upqXM6-1YwKrvUAGJkLDsTvI1GDB29uaMLPeb2ejUmP3eydOunTinacz+VnBIA */
-  context: ({ input }) => {
-    return {
-      userAddress: input.userAddress,
-      userChainType: input.userChainType,
-      chain: input.chain,
-      depositAddress: null,
-      error: null,
-    }
+  /** @xstate-layout N4IgpgJg5mDOIC5QTABwPawJYBcDiYAdmAE4CGOYAghBCXLALJkDGAFlsQGoBMAxACUAogEUAqkIDKAFQD6eIQDkhAqtKGyqAES3DJkgNoAGALqJQGbDizpC5kAA9EPAIwBWAHQuXAFgDsAJz+fu6hAGwANCAAnogAHGEePEYAzAFhbq5xKWlxbgC++VEolrgExOSUNHQMzOycYLyCohIy8koqahraulKGLmZIIKXWtvZOCG5xcR5xATx+ya4hcTxuUbEIWR5+YQEpfj4pa36ZBUUgJZhlRKQU1LT0sEysHNw8HjAVFJxQfBC2MAeTgAN3QAGsgVcrOU7lVHrVXg1eJ9bpVfghQegWD9bMYTPj7CMbHYhhMUkZEnMAvs3LsfGEUmF3BtECk6R4plk3DkXMy3HzCsU0Nd8Gj7tUni96u9Ud9rIQ-qQSOgSB5UAAbCgAM1VAFsPNCbvKHjVnnU3o0Pl84RisTjRoR8YShsSxmS2UY-JywoFkjyfDS-MHWQgfD4jF4eEc3GE4kcEu5ChdCOgUPAhkaxSbJYiZVaiaKSeNEHzvd5-EFg6EBZEYogALQuJLpVJGdJ7MI8abxoWXEUw8Xws3Sy0om3oxWFqzFj1bfYeFLxlIuA5+dlc0M8A4eHzeEJpL3dtJ9rOwyqmqUW5EfFjoPWasCUCDT3Cz0DkvyRoxGHn0nw8HsASUqGdIzDkRguAevopBGPCngOxpwpeeZjh8WAQBqYCvo6JYIIcAS7jGkEMisayhkESTRi4RhLLRjK+sm+RAA */
+  context: {
+    userAddress: null,
+    userChainType: null,
+    blockchain: null,
+    preparationOutput: null,
   },
 
-  id: "deposit-generate-address",
+  id: "depositGenerateAddressMachine",
 
-  initial: "generating",
-
-  output: ({ context }): Output => {
-    if (context.depositAddress != null) {
-      return {
-        tag: "ok",
-        value: {
-          depositAddress: context.depositAddress,
-          depositDescription: {
-            type: "depositAddressGenerated",
-            depositAddress: context.depositAddress,
-          },
-        },
-      }
-    }
-
-    if (context.error != null) {
-      return context.error
-    }
-
-    throw new Error("Unexpected output")
+  on: {
+    REQUEST_GENERATE_ADDRESS: [
+      {
+        actions: ["resetPreparationOutput", "setInputParams"],
+        target: ".generating",
+        guard: "isInputSufficient",
+      },
+      ".completed",
+    ],
   },
-
-  description:
-    "Handle initializing a new deposit procedure.\n\nResult \\[Branching\\]:\n\n1. Network base direct deposit (Near, Ethereum, etc.);\n2. Indirect deposit via POA bridge;",
 
   states: {
     generating: {
       invoke: {
         input: ({ context }) => {
-          if (!context.chain || !context.userAddress) {
-            throw new Error("Asset address or account ID is missing")
-          }
+          assert(context.userAddress, "userAddress is null")
+          assert(context.userChainType, "userChainType is null")
+          assert(context.blockchain, "blockchain is null")
           return {
             userAddress: context.userAddress,
             userChainType: context.userChainType,
-            chain: context.chain,
+            blockchain: context.blockchain,
           }
         },
         onDone: {
-          target: "Genereted",
+          target: "completed",
           actions: assign({
-            depositAddress: ({ event }) => event.output,
+            preparationOutput: ({ event }) => ({
+              tag: "ok",
+              value: {
+                generateDepositAddress: event.output,
+              },
+            }),
           }),
+
+          reenter: true,
         },
         onError: {
-          target: "Failed",
-          actions: [
-            {
-              type: "logError",
-              params: ({ event }) => event,
-            },
-            {
-              type: "setError",
-              params: ({ event }) => {
-                return {
-                  reason: "ERR_GENERATING_ADDRESS",
-                  error: toError(event.error),
-                }
+          target: "completed",
+          actions: assign({
+            preparationOutput: {
+              tag: "err",
+              value: {
+                reason: "ERR_GENERATING_ADDRESS",
               },
             },
-          ],
+          }),
+
+          reenter: true,
         },
         src: "generateDepositAddress",
       },
-      description: "Generate deposit address via POA bridge",
     },
-    Genereted: {
-      type: "final",
-    },
-    Failed: {
-      type: "final",
-    },
-  },
-})
 
-function toError(error: unknown): Error {
-  return error instanceof Error ? error : new Error("unknown error")
-}
+    completed: {},
+    idle: {},
+  },
+
+  initial: "idle",
+})
